Updates to accept decline
trying to make a baseline file that can accept / decline things and have
variable commands associated with it, but also just have basic "accept"
or "deny" command. I want it to be versatile and simple. And never have
to touch it again. I still have to make some updates to how commands are
featured.
I want to be able to add the class on the fly and use it.